requirements.txt
requests~=2.31.0
icecream~=2.1.3
pymongo~=4.6.0
numpy~=1.26.2
python-tlsh
ssdeep